backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
phonenumberslite==8.13.22
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
